// src/integration/exchange/dto/exchange-tx-transaction.dto.ts
import { ExchangeTxType } from '../entities/exchange-tx.entity';

export class ExchangeTxTransactionDto {
  exchange: string;
  type: ExchangeTxType;
  externalId: string;
  externalCreated?: Date;
  externalUpdated?: Date;
  status?: string;
  originalStatus?: string;
  amount?: number;
  cost?: number;
  currency?: string;
  feeAmount?: number;
  feeCurrency?: string;
  method?: string;
  aClass?: string;
  asset?: string;
  network?: string;
  address?: string;
  addressTo?: string;
  addressFrom?: string;
  refId?: string;
  txId?: string;
  tag?: string;
  tagTo?: string;
  tagFrom?: string;
  order?: string;
  orderTxId?: string;
  posTxId?: string;
  pair?: string;
  orderType?: string;
  price?: number;
  vol?: number;
  margin?: number;
  leverage?: number;
  misc?: string;
  tradeId?: string;
  symbol?: string;
  side?: string;
  takeOrMaker?: string;
}